+++ NEWS.md
@@ -2,3 +2,5 @@
 
 * Labels created with the `labs` argument to `ggtibble()` will not longer all be
   the same (#3)
+* `new_gglist()` and `new_ggtibble()` are now exported making it easier to
+  create objects.